+++ testing.py
@@ -106,7 +106,6 @@ for server in servers:
 
         for site in sites:
 
-            
             siteStartTime = datetime.now()
             
             coordinates = site['geometry']['coordinates']
@@ -223,10 +222,8 @@ for server in servers:
 
                         if numberBasinCharacteristicsNotEqualToKnownValues == 0:
                             printOut("Completed. All computed values were equal to known values.")
-                        elif numberBasinCharacteristicsNotEqualToKnownValues == 1:
-                            printOut("Completed. 1 value was not equal to known values.")
                         else:
-                            printOut("Completed. " + str(numberBasinCharacteristicsNotEqualToKnownValues) + " values were not equal to known values.")
+                            printOut("Completed. Number of Basin Characteristics not equal to known values: " + str(numberBasinCharacteristicsNotEqualToKnownValues))
 
                     except Exception as e:
                         printOut("Failed. Retrying...")
@@ -300,7 +297,240 @@ for server in servers:
     except Exception as e:
         # print(e)
         print(server.upper() + " server is down. Moving on to next server.")
-    
+
+## Compare Test to ProdWebA and ProdWebB
+
+# All comparision results files will go in Output/Testing-YYYY-MM-DD-HH-MM-SS/Comparision directory
+comparisonDirectory = os.path.join(dateDirectory, "Comparison")
+if not os.path.exists(comparisonDirectory):
+    os.makedirs(comparisonDirectory)
+
+# Create a summary file to summarize differences 
+fileName = os.path.join(comparisonDirectory, "DifferencesSummary.csv")
+comparisonSummaryFile = open(fileName, "w", newline='')
+comparisonSummaryFileWriter = csv.writer(comparisonSummaryFile)
+headerRow = ["Testing session", "Server 1", "Server 2", "Region", "SiteID", "Task", "Description", "Server 1 Value", "Server 2 Value"]
+comparisonSummaryFileWriter.writerow(headerRow)
+comparisonSummaryFile.flush()
+
+productionServers = ["PRODWEBA", "PRODWEBB"]
+
+# Compare Test to ProdWebA or ProdWebB
+for productionServer in productionServers:
+
+    # Results files for this server comparision will go in Output/Testing-YYYY-MM-DD-HH-MM-SS/Comparision/Comparison-TEST-[PRODWEBA OR PRODWEBB] directory
+    comparisonTestProdDirectory = os.path.join(comparisonDirectory, "Comparison-TEST-" + productionServer)
+    if not os.path.exists(comparisonTestProdDirectory):
+        os.makedirs(comparisonTestProdDirectory)
+
+    testDirectory = os.path.join(dateDirectory, "TEST")
+    prodDirectory = os.path.join(dateDirectory, productionServer)
+
+    ## Create files where comparisons will be saved
+    # This file will include all comparisions
+    fileName = os.path.join(comparisonTestProdDirectory, "Comparison-TEST-" + productionServer + ".csv")
+    comparisonTestProdFile = open(fileName, "w", newline='')
+    comparisonTestProdFileWriter = csv.writer(comparisonTestProdFile)
+    headerRow = ["Testing session", "Server 1", "Server 2", "Region", "SiteID", "Task", "Description", "Server 1 Value", "Server 2 Value", "Server 1 value equal to Server 2 value?"]
+    comparisonTestProdFileWriter.writerow(headerRow)
+    comparisonTestProdFile.flush()
+
+    # This file will include only results that were different between the 2 servers
+    fileName = os.path.join(comparisonTestProdDirectory, "Comparison-TEST-" + productionServer + "-Differences.csv")
+    comparisonTestProdDifferencesFile = open(fileName, "w", newline='')
+    comparisonTestProdDifferencesFileWriter = csv.writer(comparisonTestProdDifferencesFile)
+    headerRow = ["Testing session", "Server 1", "Server 2", "Region", "SiteID", "Task", "Description", "Server 1 Value", "Server 2 Value"]
+    comparisonTestProdDifferencesFileWriter.writerow(headerRow)
+    comparisonTestProdDifferencesFile.flush()
+
+    # This fill will include only results that were not able to be compared because they were unavailable on one server
+    fileName = os.path.join(comparisonTestProdDirectory, "Comparison-TEST-" + productionServer + "-Uncompared.csv")
+    comparisonTestProdUncomparedFile = open(fileName, "w", newline='')
+    comparisonTestProdUncomparedFileWriter = csv.writer(comparisonTestProdUncomparedFile)
+    headerRow = ["Testing session", "Server 1", "Server 2", "Region", "SiteID", "Task", "Description", "Server 1 Value", "Server 2 Value"]
+    comparisonTestProdUncomparedFileWriter.writerow(headerRow)
+    comparisonTestProdUncomparedFile.flush()
+
+    printOut("========== COMPARING TEST AND " + productionServer + " RESULTS ==========")
+
+    for site in sites:
+
+        properties = site['properties']
+        siteID = properties['siteid']
+        region = properties['state']
+
+        printOut("*** " + region + " ***")
+
+        # Compare Basin Delineation results
+        printOut("BASIN DELINEATION:")
+
+        try:
+            testBasinDelineationDirectory = os.path.join(testDirectory, "BasinDelineations")
+            testBasinDelineationsFileName = os.path.join(testBasinDelineationDirectory, region + ".txt")
+            testBasinDelineationsFileNotFound = False        
+            testBasinDelineationsFile = open(testBasinDelineationsFileName, "r")
+        except FileNotFoundError:
+            testBasinDelineationsFileNotFound = True
+
+        try:
+            prodBasinDelineationDirectory = os.path.join(prodDirectory, "BasinDelineations")
+            prodBasinDelineationsFileName = os.path.join(prodBasinDelineationDirectory, region + ".txt")
+            prodBasinDelineationsFileNotFound = False
+            prodBasinDelineationsFile = open(prodBasinDelineationsFileName, "r")
+        except FileNotFoundError:
+            prodBasinDelineationsFileNotFound = True
+        
+        if (testBasinDelineationsFileNotFound is False and prodBasinDelineationsFileNotFound is False):
+            testBasinDelineationCoordinates = json.load(testBasinDelineationsFile)['featurecollection'][1]['feature']['features'][0]["geometry"]['coordinates']
+            prodBasinDelineationCoordinates = json.load(prodBasinDelineationsFile)['featurecollection'][1]['feature']['features'][0]["geometry"]['coordinates']
+            if testBasinDelineationCoordinates == prodBasinDelineationCoordinates:
+                printOut("Basin Delineations are equal.")
+            else: 
+                printOut("Basin Delineations are not equal.")
+                dataRow = [folderName, "TEST", productionServer, region, siteID, "BasinDelineation", "", "See file", "See file"]
+                comparisonTestProdDifferencesFileWriter.writerow(dataRow)
+                comparisonTestProdDifferencesFile.flush()
+                comparisonSummaryFileWriter.writerow(dataRow)
+                comparisonSummaryFile.flush()
+            dataRow = [folderName, "TEST", productionServer, region, siteID, "BasinDelineation", "", "See file", "See file", str(testBasinDelineationCoordinates == prodBasinDelineationCoordinates)]
+            comparisonTestProdFileWriter.writerow(dataRow)
+            comparisonTestProdFile.flush()
+        elif (testBasinDelineationsFileNotFound is True and prodBasinDelineationsFileNotFound is True):
+            printOut("Basin Delineations are equal.")
+            dataRow = [folderName, "TEST", productionServer, region, siteID, "BasinDelineation", "", "See file", "See file", str(True)]
+            comparisonTestProdFileWriter.writerow(dataRow)
+            comparisonTestProdFile.flush()
+        else:
+            printOut("Basin Delineations available for only one server. Cannot compare.")
+            dataRow = [folderName, "TEST", productionServer, region, siteID, "BasinDelineation", "", "File not found" if testBasinDelineationsFileNotFound else "See file", "File not found" if prodBasinDelineationsFileNotFound else "See file", "Not applicable"]
+            comparisonTestProdFileWriter.writerow(dataRow)
+            comparisonTestProdFile.flush()
+            dataRow = [folderName, "TEST", productionServer, region, siteID, "BasinDelineation", "", "File not found" if testBasinDelineationsFileNotFound else "See file", "File not found" if prodBasinDelineationsFileNotFound else "See file"]
+            comparisonTestProdUncomparedFileWriter.writerow(dataRow)
+            comparisonTestProdUncomparedFile.flush()
+            comparisonSummaryFileWriter.writerow(dataRow)
+            comparisonSummaryFile.flush()
+
+        # Compare Basin Characteristics results
+        printOut("BASIN CHARACTERISTICS:")
+        numberBasinCharacteristicsNotEqual = 0
+
+        try:
+            testBasinCharacteristicsDirectory = os.path.join(testDirectory, "BasinCharacteristics")
+            testBasinCharacteristicsFileName = os.path.join(testBasinCharacteristicsDirectory, region + ".txt")
+            testBasinCharacteristicsFileNotFound = False
+            testBasinCharacteristicsFile = open(testBasinCharacteristicsFileName, "r")
+        except FileNotFoundError:
+            testBasinCharacteristicsFileNotFound = True
+
+        try:
+            prodBasinCharacteristicsDirectory = os.path.join(prodDirectory, "BasinCharacteristics")
+            prodBasinCharacteristicsFileName = os.path.join(prodBasinCharacteristicsDirectory, region + ".txt")
+            prodBasinCharacteristicsFileNotFound = False
+            prodBasinCharacteristicsFile = open(prodBasinCharacteristicsFileName, "r")
+        except FileNotFoundError:
+            prodBasinCharacteristicsFileNotFound = True
+        
+        if (testBasinCharacteristicsFileNotFound is False and prodBasinCharacteristicsFileNotFound is False):
+            # Compare the files
+            testParameters = json.load(testBasinCharacteristicsFile)['parameters']
+            testDictionary = {}
+            for testParameter in testParameters:
+                testDictionary[testParameter['code']] = testParameter['value']
+            prodParameters = json.load(prodBasinCharacteristicsFile)['parameters']
+            for prodParameter in prodParameters:
+                if testDictionary[prodParameter['code']] != prodParameter['value']:
+                    printOut("Basin Characteristic not equal: " + prodParameter['code'])
+                    printOut("TEST value = " + str(testDictionary[prodParameter['code']]))
+                    printOut(productionServer + " value = " + str(prodParameter['value']))
+                    numberBasinCharacteristicsNotEqual+= 1
+                    dataRow = [folderName, "TEST", productionServer, region, siteID, "BasinCharacteristics", prodParameter['code'], str(testDictionary[prodParameter['code']]), str(prodParameter['value'])]
+                    comparisonTestProdDifferencesFileWriter.writerow(dataRow)
+                    comparisonTestProdDifferencesFile.flush()
+                    comparisonSummaryFileWriter.writerow(dataRow)
+                    comparisonSummaryFile.flush()
+                dataRow = [folderName, "TEST", productionServer, region, siteID, "BasinCharacteristics", prodParameter['code'], str(testDictionary[prodParameter['code']]), str(prodParameter['value']), str(testDictionary[prodParameter['code']] == prodParameter['value'])]
+                comparisonTestProdFileWriter.writerow(dataRow)
+                comparisonTestProdFile.flush()
+            if numberBasinCharacteristicsNotEqual > 0:
+                printOut("Number of Basin Characteristics not equal between servers: " + str(numberBasinCharacteristicsNotEqual))
+            else:
+                printOut("All Basin Characteristics are equal.")
+        elif (testBasinCharacteristicsFileNotFound is True and prodBasinCharacteristicsFileNotFound is True):
+            printOut("All Basin Characteristics are equal.")
+            dataRow = [folderName, "TEST", productionServer, region, siteID, "BasinCharacteristics", "", "File not found", "File not found"]
+            comparisonTestProdFileWriter.writerow(dataRow)
+            comparisonTestProdFile.flush()
+        else:
+            printOut("Basin Characteristics available for only one server. Cannot compare.")
+            dataRow = [folderName, "TEST", productionServer, region, siteID, "BasinCharacteristics", "", "File not found" if testBasinCharacteristicsFileNotFound else "See file", "File not found" if prodBasinCharacteristicsFileNotFound else "See file"]
+            comparisonTestProdUncomparedFileWriter.writerow(dataRow)
+            comparisonTestProdUncomparedFile.flush()
+            comparisonSummaryFileWriter.writerow(dataRow)
+            comparisonSummaryFile.flush()
+
+
+        # Compare Flow Statistics results
+        printOut("FLOW STATISTICS:")
+        numberFlowStatisticsNotEqual = 0
+
+        try:
+            testFlowStatisticsDirectory = os.path.join(testDirectory, "FlowStatistics")
+            testFlowStatisticsFileName = os.path.join(testFlowStatisticsDirectory, region + ".txt")
+            testFlowStatisticsFileNotFound = False
+            testFlowStatisticsFile = open(testFlowStatisticsFileName, "r")
+        except FileNotFoundError:
+            testFlowStatisticsFileNotFound = True
+
+        try:
+            prodFlowStatisticsDirectory = os.path.join(prodDirectory, "FlowStatistics")
+            prodFlowStatisticsFileName = os.path.join(prodFlowStatisticsDirectory, region + ".txt")
+            prodFlowStatisticsFileNotFound = False
+            prodFlowStatisticsFile = open(prodFlowStatisticsFileName, "r")
+        except FileNotFoundError:
+            prodFlowStatisticsFileNotFound = True
+        
+        if (testFlowStatisticsFileNotFound is False and prodFlowStatisticsFileNotFound is False):
+            # Compare the files
+            testStatisticGroups = json.load(testFlowStatisticsFile)
+            testDictionary = {}
+            for testStatisticGroup in testStatisticGroups:
+                testDictionary[testStatisticGroup['StatisticGroupID']] = {}
+                for testRegressionRegion in testStatisticGroup['RegressionRegions']:
+                    testDictionary[testStatisticGroup['StatisticGroupID']][testRegressionRegion['Code']] = {}
+                    for testResult in testRegressionRegion["Results"]:
+                        testDictionary[testStatisticGroup['StatisticGroupID']][testRegressionRegion['Code']][testResult["code"]] = testResult["Value"]
+
+            prodStatisticGroups = json.load(prodFlowStatisticsFile)
+            for prodStatisticGroup in prodStatisticGroups:
+                for prodRegressionRegion in prodStatisticGroup['RegressionRegions']:
+                    for prodResult in prodRegressionRegion["Results"]:
+                        if testDictionary[prodStatisticGroup['StatisticGroupID']][prodRegressionRegion['Code']][prodResult["code"]] != prodResult["Value"]:
+                            numberFlowStatisticsNotEqual += 1
+                            dataRow = [folderName, "TEST", productionServer, region, siteID, "FlowStatistics", prodResult["code"], str(testDictionary[prodStatisticGroup['StatisticGroupID']][prodRegressionRegion['Code']][prodResult["code"]]), str(prodResult["Value"])]
+                            comparisonTestProdDifferencesFileWriter.writerow(dataRow)
+                            comparisonTestProdDifferencesFile.flush()
+                            comparisonSummaryFileWriter.writerow(dataRow)
+                            comparisonSummaryFile.flush()
+                        dataRow = [folderName, "TEST", productionServer, region, siteID, "FlowStatistics", prodResult["code"], str(testDictionary[prodStatisticGroup['StatisticGroupID']][prodRegressionRegion['Code']][prodResult["code"]]), str(prodResult["Value"]), str(testDictionary[prodStatisticGroup['StatisticGroupID']][prodRegressionRegion['Code']][prodResult["code"]] == prodResult["Value"])]
+                        comparisonTestProdFileWriter.writerow(dataRow)
+                        comparisonTestProdFile.flush()
+            if numberFlowStatisticsNotEqual > 0:
+                printOut("Number of Flow Statistics not equal: " + str(numberFlowStatisticsNotEqual))
+            else:
+                printOut("All Flow Statistics are equal.")
+        elif (testFlowStatisticsFileNotFound is True and prodFlowStatisticsFileNotFound is True):
+            printOut("Flow Statistics are equal.")
+            dataRow = [folderName, "TEST", productionServer, region, siteID, "FlowStatistics", "", "File not found", "File not found", str(True)]
+            comparisonTestProdFileWriter.writerow(dataRow)
+            comparisonTestProdFile.flush()
+        else:
+            printOut("Flow Statistics available for only one server. Cannot compare.")
+            dataRow = [folderName, "TEST", productionServer, region, siteID, "FlowStatistics", "", "File not found" if testFlowStatisticsFileNotFound else "See file", "File not found" if prodFlowStatisticsFileNotFound else "See file"]
+            comparisonTestProdUncomparedFileWriter.writerow(dataRow)
+            comparisonTestProdUncomparedFile.flush()
+            comparisonSummaryFileWriter.writerow(dataRow)
+            comparisonSummaryFile.flush()
 
 overallEndTime = datetime.now()
 overallTimeElapsed = overallEndTime - overallStartTime